requirements.txt
flask==2.3.2
pydantic==1.10.9
openai==0.27.0
requests==2.31.0
gunicorn==20.1.0
python-dotenv==1.0.0
werkzeug>=2.3.3
